+++ migrations/sqls/20220324021344-order-products-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE order_products (
     id SERIAL PRIMARY KEY NOT NULL,
-    order_id bigint REFERENCES orders(id),
-    product_id bigint REFERENCES products(id),
+    order_id INTEGER REFERENCES orders(id),
+    product_id INTEGER REFERENCES products(id),
     quantity INTEGER
 );
